lightbox: Replace onDismiss with `await showErrorDialog`
The `showDialog` function returns a future that resolves when the
dialog is dismissed, and showErrorDialog returns that future.

The one caller of showErrorDialog that was passing an onDismiss
callback can therefore await that future instead.  Do that, and
eliminate the parameter.

This is not an NFC change, because now the user can tap the barrier
(the darkened area around the dialog) to close it.

Discussion from when this parameter was introduced:
  https://github.com/zulip/zulip-flutter/pull/587#discussion_r1585845428

Signed-off-by: Zixuan James Li <zixuan@zulip.com>