check [nfc]: Say `git diff -a` rather than `git diff --text`
Normally in scripts I prefer using long options, to make
the meaning more explicit for the reader.

But this particular option `--text` on `git diff` has a pretty
confusing name: it sounds like it potentially means "only look
at text files", which is the opposite of what it turns out to mean.
Plus its meaning really isn't about "text" (vs "binary") at all: the
files we set the `-diff` attribute on are mostly just as textual as
any others, just boring for diffs because they're generated.

So use the short name `-a` instead.  That suggests "all", which is
exactly the right idea.

Also expand the comment a bit.