subscription_list: Place ad hoc dark-theme color for MutedUnreadBadge
Thanks to Greg for the reminder that this needed attention:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/dark.20theme/near/1936853

Web, in `main`, uses hsl(0, 0%, 30%) with 50% opacity. We tried
using that, but it ended up with pretty low contrast, given the
different background it's on here in mobile. So, this ad hoc value,
until we have a design from Vlad.

Web's light and dark colors might both actually change soon, with
the in-flight PR zulip/zulip#31847 that's test-deployed on CZO. But
not by much, and not enough to fix the low-contrast issue for us in
dark theme. I think the intended change is just to bump the opacity
from 50% to 55%. So we might refresh the light variant after that PR
is merged, and in any case will refresh both once we have Vlad's
design.