emoji [nfc]: Move "how to show this emoji" logic to EmojiStore
This will apply in other places we show an emoji, too,
such as the emoji picker.

What, you might ask, about the other place we already show emoji,
namely in message content?  We should probably apply this logic
there too, but we'll leave that for later.  Tracking it as issue #966.