deps: Upgrade video_player_android to 2.7.5
Changelog:
  https://pub.dev/packages/video_player_android/changelog

This update addresses a few bug fixes, including the
suppression of deprecation warnings for `onSurfaceCreated`,
which was introduced in an early version of the
Impeller-compatible Texture APIs. For more info, see note of
the migration guide:
  https://docs.flutter.dev/release/breaking-changes/android-surface-plugins#migration-guide

Without this update, Android builds will fail with Flutter SDK
versions after the following change that landed yesterday:
  https://github.com/flutter/engine/pull/55463
  https://github.com/flutter/flutter/commit/7bb9352911d0471568a77fc57b94a646fe9571af
due to our strict warnings as errors config in `app/build.gradle`.
The build error would look like this:

  /…/video_player_android-2.7.3/android/src/main/java/io/flutter/plugins/videoplayer/VideoPlayer.java:88:
  warning: [removal] onSurfaceCreated() in Callback has been deprecated and marked for removal
    public void onSurfaceCreated() {
                ^
  error: warnings found and -Werror specified