store: Ignore boring NetworkExceptions
Usually when a client goes back from sleep or lose network connection,
there will be a bunch of errors when polling event queue.  This known
error will always be a `NetworkException`, so we make a separate case
for it.

Previously, we may report these errors, and it ended up being spammy.
This filters out the more interesting one to report instead.

Fixes: #555

Signed-off-by: Zixuan James Li <zixuan@zulip.com>