test: Add await for TestGlobalStore.add calls
These tests didn't break previously because the GlobalStore.add calls
finished before the per account store was ever needed.

A notable example is the test "find account among several" in
test/notifications/display_test.dart.  The accounts were added in
parallel, which actually allowed the GlobalStore.add calls to bypass
the invariant check performed by TestGlobalStore.  The test passed
anyway because it only requires the accounts to be there when some
other checks are performed.

The invariant check is to ensure that no accounts have the same email
address within a realm.  To fix this, we can potentially maintain a
store of accounts with TestGlobalStore, and check for the constraints
there.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>