store test: Add finish for checking Future.throws
These tests actually worked as expected, even though we didn't await
for the Future returned by FutureChecks.throws.

This is because checks ensure that the asychronous expectations are
complete before the test ends.

Nonetheless, we should use finish so that our intention is explicit
(i.e. the test runner should wait for the future to complete).

See also:
  https://github.com/dart-lang/test/tree/master/pkgs/checks#checking-asynchronous-expectations

Signed-off-by: Zixuan James Li <zixuan@zulip.com>