backoff [nfc]: Cut _startTime field, never used
This field is initialized but never looked at thereafter.
So we can just cut it.

It looks like it was never used in the previous history of this
file, nor in the history of the corresponding code in zulip-mobile.
Presumably it was left over from an earlier draft of that
zulip-mobile implementation.