backoff [nfc]: Use Dart Duration values more thoroughly
This code was originally adapted from an implementation in JavaScript
(from zulip-mobile), and followed that implementation in representing
durations mostly as doubles in milliseconds.

Before making other changes here, and in particular exposing some
of these durations in the class's API, let's adapt it more fully to
Dart idioms and norms.