backoff test [nfc]: Fix a long `=>` expression body, preferring block body
The use of an expression body (with `=>`) in this situation is
forbidden in the upstream Flutter style guide:
  https://github.com/flutter/flutter/blob/e6c955416da/docs/contributing/Style-guide-for-Flutter-repo.md#consider-using--for-short-functions-and-methods
  https://github.com/flutter/flutter/blob/e6c955416da/docs/contributing/Style-guide-for-Flutter-repo.md#use--for-getters-and-callbacks-that-just-return-literals-or-switch-expressions

and I've regularly asked in code reviews to replace such things
with block bodies.

Apparently I hadn't yet absorbed that point, though, when I wrote
this code earlier this year.  And then just now I was rereading it
in order to refactor it, and found I had to visually scan the test's
code several times in order to see its overall structure.  Fix that.