poll: Move vote count paddings inside children of Row
It makes it easier to adjust the size of the vote count box touch
target later.

This preserves most of the visuals. Except that when the vote count
box is shorter than the voter names text, the padding between rows
is gone.

The end padding is equivalent to the spacing previously specified on
the Row; the bottom padding is (mostly) equivalent to the padding
specified on the outer (now removed) Padding.  The ConstrainedBox is
expanded because it now also contains the paddings.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>