poll [nfc]: Move vote count paddings inside children of Row
It makes it easier to adjust the size of the vote count box touch
target later.

This preserves the visuals. In particular, we add bottom paddings
to both the vote count box and the option/voters text to ensure
that there is always 5px between option rows, as it was before.

The end padding is equivalent to the spacing previously specified on
the Row; the bottom padding is equivalent to the padding specified on
the outer (now removed) Padding.  The ConstrainedBox is expanded
because it now also contains the paddings.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>