poll [nfc]: Move vote count paddings to a box
While this mostly preserves the visuals, there might be some subtle
changes the paddings.  It makes it easier to adjust the size of the
vote count box touch target later.

The end padding is equivalent to the spacing previously specified on
the Row; the bottom padding is equivalent to the padding specified on
the outer (now removed) Padding.  The Container is expanded because it
now also contains the paddings.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>