deps: Upgrade video_player_android to 2.7.8
Changelog:
  https://pub.dev/packages/video_player_android/changelog

This update addresses a few bug fixes, including the removal of a flag
that treats warnings as errors. Without this fix, Android builds fail
in Android Studio 2024.2, which uses Java 21, with the
following error:
  warning: [options] source value 8 is obsolete and will be removed in a future release
  warning: [options] target value 8 is obsolete and will be removed in a future release
  warning: [options] To suppress warnings about obsolete options, use -Xlint:-options.
  error: warnings found and -Werror specified

With this update the warnings still stay, but are ignored by build
system resulting in a successful build.

[Greg adds: the changelog shows one potentially user-facing bug fix,
when reopening the app from the background while a video was playing;
but I couldn't reproduce that issue in our app.]