poll: Enlarge and vetically center vote count box
Previously, there were only bottom and end paddings; we redistribute
the vertical paddings so the vote count box is padded on both sides.

The end padding remains 5px because we want to keep the options aligned
to the start with the surrounding messages.

The inner padding defined on the Container is affected by
the size of the text.  Document that fact as we make the green box
and the text in it larger.

This change also affects the poll question, where 2.5px of space gets
moved to the first option row.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>