dialog [nfc]: Simplify new showErrorDialog interface a bit
Unlike ScaffoldFeatureController which inspired it, this class
doesn't offer a way to control the dialog, so it's not really
a "controller"; just call it DialogStatus.

A private final field with a public getter (of the same type)
is equivalent to just a public final field: nothing can set it,
and any library can get it.

Also generalize the description a bit: we currently use this
only for AlertDialog, but it'd apply equally well for any
dialog box.