compose_box: Send "typing stopped" notices on app lifecycle updates
This is a bonus feature that covers some cases that FocusNode doesn't
cover.  We send a "typing stopped" notice when the app loses focus or
becomes invisible.

An example of WidgetsBindingObserver can be found here:
  https://api.flutter.dev/flutter/widgets/WidgetsBindingObserver-class.html

The cited text comes from the AppLifecycleState documentation:
  https://github.com/flutter/engine/blob/a65f1d59edc618ae81e2e8ed78d59fb729291afa/lib/ui/platform_dispatcher.dart#L1856-L1991

The link is not included in code because the code themselves are
references to their documentation already.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>