doc: Add section on getting help with codebase
Apparently we didn't actually say anywhere in these docs
where one should go to ask for help with contributing!
(Though we did link to some Zulip-wide resources that included
links to the right channel.)

Fix that.  And take the opportunity to stress the community norms;
we often see new participants who ask for help that clearly
haven't read them.