api [nfc]: Require `appid` for addApnsToken route
This is required for current servers, since server-8.

The original binding for this route didn't mark it as required
because it was based on the reverse-engineered binding in
zulip-mobile, and that one hadn't been updated for server-8.
Now that the endpoint has API docs, we can use the information
they tell us.