store: Call connection.close in PerAccountStore.dispose
This seems hygienic. It makes sense to explicitly clean up an old
HTTP client's resources when we've set up a new one, for the
dead-queue reload. (The dead-queue reload is currently the only path
where this dispose method is called.)

It will also make sense to clean up these resources when an account
is logged out. That's a feature we'll be implementing soon, with
PerAccountStore.dispose in that new path as well.

The "abort long-poll and close ApiConnection" TODO was on
UpdateMachine's dispose method, not PerAccountStore's. But since
PerAccountStore is what owns the connection, its dispose method
seemed like the more appropriate place to close the connection.