lint: Re-enable analyzer on Pigeon-generated code
The upstream issue that made this exclusion necessary has been fixed:
  https://github.com/flutter/flutter/issues/145633

In fact it turns out it was fixed in pigeon 18.0.0, which we took
in 8df306412 back in April; we just didn't notice at the time.

Meanwhile this file grew a different analyzer issue during the time it
wasn't getting checked -- and that one was an actual actionable one
for us to fix, where a doc comment still referred to a deprecated bit
of API.  Fixed that in the previous commit.  Glad to have the analyzer
re-enabled here.