compose [nfc]: Refactor away shared iconTheme override
Because these buttons share the same superclass, we can de-duplicate
styling code without the theme override.  This helps de-nest the
layout code.

While more types of buttons that are not subclasses of
`_AttachUploadsButton` might be added, a theme is probably still
unnecessary given that we can lift a common button class that handles
the styling.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>