dialog: Dismiss suggested-action dialog on pressing action button
So far this function just has two callers in uncommon paths
involving app permissions. For those, and for the new caller we're
about to add -- logging out an account -- it's unhelpful for the
dialog to linger unchanged after the action button is tapped. Better
to make it go away, preventing confusion about whether the action
was actually done.