compose_box test: Set topic and content through a helper
This ensures that the `GetStreamTopicsResult` is self-contained, so
that the caller does not rely on the assumption that a response has
been prepared through `prepareComposeBox`.

This also elevates contentInputFinder for reuse throughout the test.

This changes the way the tests work, because now we use tester.enterText
to set the topic and content, instead of accessing the controllers.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>