msglist [nfc]: Remove no-op Center widget in _MessageListPageState
The Center's child is a Column with the message list and possibly a
compose box. That Column expresses that it wants to fill all the
available space (which I agree seems desirable), so there's only one
position it can have in that space. Asking it to be centered there
is a harmless no-op but is confusing when reading the code, so, stop
doing so.

This Center looks like a relic from the very early prototype in late
2022; see commit b22ef6752.