msglist [nfc]: Remove no-op Center widget in _MessageListPageState
The Center's child is a Column with the message list and possibly a
compose box.

Since the Column has the default MainAxisSize.max, it occupies all
available vertical space, so the Center is a no-op on the vertical
axis.

(If an element occupies all available space on an axis, it has only
one possible position on that axis, so the Center can't affect its
position.)

It turns out that the Column always takes all horizontal space, too;
the message-list and compose-box widgets both stretch to the full
screen width. This is appropriate because they need to handle the
horizontal device insets in different ways: the former by keeping
visible UI outside the insets; the latter by filling the insets with
padding. So the Center is a no-op in the horizontal direction, too.

It's a relic from the very early prototype in late 2022 (see commit
b22ef6752); remove it.

Also, with some new comments and dartdoc, make it clearer that the
Column takes the full screen width.