msglist [nfc]: Explain an odd-looking Center widget that we want to keep
Looks like I explained the Center widget in the commit message that
added it -- ef02ee4e8 -- but I think the comment is helpful. At
first, it looked like another case of a no-op Center widget like in
the previous commit, but it turns out it does have a desirable
effect in the horizontal axis.