compose_box test: Enter topic with a helper
A request looking up topics within a channel is sent whenever the
topic input gets edited.  Previously, the response to this request was
prepared in a general helper.  We move it to a more focused helper to
improve reusability, and to simplify the general helper, because only
two test cases currently need this feature.

This also includes a brief check that consumes the reqeust, confirming
that the request is indeed the expected one.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>