compose_box: Send typing notices on content/focus changes
For a compose box with topic input, the field `destination`
contains a mutable state computed from the topic text.

This implementation is similar to hintText's to still keep
everything up-to-date.

When testing typing activities that do not end with a "typing stopped"
notice, we need to wait for the idle timer to expire so that the test
does not end with pending timers.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>