action_sheet [nfc]: Tighten button constructors onto one line
This is a nice small bonus from renaming the parameter that was
`messageListContext` to the shorter name `pageContext`.

Some of these don't quite fit in 80 columns.  But the details of
the parameter list are boring (because they're all the same),
so it's OK that some of them go a little over.