test store: Make TestGlobalStore.doInsertAccount not drop ackedPushToken
Without this line, if a test `globalStore.add`s an account with a
non-null `ackedPushToken`, the field will be null in app code or
test code that retrieves the account from `globalStore`. Thankfully
no tests have been doing that, but it'll be convenient to do that
soon, for testing the unregister-token part of logging out.