store [nfc]: Move reload-store logic to enclosing whole poll method
As is, this structure looks a bit silly.  But in the next few commits
we'll make use of it in order to recover from a wider range of errors
in the event-poll loop, by using the same remedy of reloading server
data and replacing the store as we use when the event queue expires.