api: Add missing includeToJson flags on some "type"/etc getters
These `JsonKey(includeToJson: true)` annotations are needed so that
the properties appear in the output of the generated toJson
implementations, like they do in the real JSON from the server.
We had the annotations on many of this kind of getter, but not all.

These types appear in responses, not requests, so the actual app
never calls these toJson methods.  But having accurate toJson methods
is handy in tests.