store [nfc]: Pull out _debugLoopWait from poll method
This takes it a bit farther in squeezing out details that aren't
part of the main picture of what this method is doing.

At this point the main remaining thing that feels like clutter
is all the `if (_disposed) return;` lines.  But I don't have any
ideas at the moment for a better way to handle those.