autocomplete test: Fix not-starve test, using awaitFakeAsync
Using just `fakeAsync`, when this hit an `await` it just stopped and
didn't finish the remainder of the test, so didn't get to the point
of testing what it's meant to test.

I believe the test worked correctly when first committed, as it had no
`await` of its own; but later was accidentally defeated by eca33f982
introducing an `await` for `store.addUsers`.

Using our `awaitFakeAsync` fixes the problem.

This was the only call to `fakeAsync` in our codebase, so I believe
this commit fixes the whole problem.