autocomplete [nfc]: Make query non-nullable on AutocompleteView
At first glance this appears to make a functional change: it causes
the search to begin as soon as the AutocompleteView is constructed,
rather than later when `query` is set to non-null.

But in fact the only times (outside tests) that we ultimately
construct an AutocompleteView are by calling an implementation of
AutocompleteField.initViewModel... and those call sites are
immediately followed by setting a non-null query.  So this is an
NFC commit after all.