compose: Disable compose box widgets when sending message
Also clear the content input only after the message is successfully
sent.

For the send button, unlike when there are validation errors, having
an error dialog when its disabled for sending a message feels a bit
redundant and doesn't help much.  This can be annoying when the user
misclicks and hit the send button multiple times.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>