theme: Apply icon color through IconButtonTheme
The most conspicuous change this makes is to the "back" icon
in an AppBar, aka AppBar.leading.  Those have been a white or
black color from Material defaults; now they get our "icon" color.

This matches what's in Figma; and also matches the action icons,
for any app bars that have both.  Currently our only app bar with
an action is also the only one that lacks a "leading": it's on
ChooseAccountPage, which is at the root of navigation so there's
nowhere for a "back" icon to lead to.  But when we add actions in
other pages, the mismatch would become conspicuous.

This change also turns a handful of less-visible icon buttons
from white/black to the icon color: the lightbox's "copy link"
or play/pause buttons, and the show/hide-password button on the
username/password login page.  I think those changes are desirable
too.  All our other icon buttons that I can find (either in the
source, or by browsing around the app) already have a color
specified, and aren't affected by this change.