api: Add ApiConnection.patch
The tests, ApiConnection.{post,patch,delete}, are mostly similar to
each other, because the majority of them are testing that the params
are parsed into the body with the same content type.

If we find the need to update these test cases with new examples, it
will be ripe to refactor them with a helper. Until then, just duplicate
them for simplicity.

Signed-off-by: Zixuan James Li <zixuan@zulip.com>