log: Report stack trace on registerQueue failure, in debug
For a malformed-response error this is essential because it pinpoints
where in the data schema the mismatch occurred between our expectations
and the server's behavior.

There's more to do here (#890 for getting this to the user in cases
like a malformed response, so they can screenshot it when reporting
an issue; #1083 for including stack traces when relevant in showing
other errors to the user) but this is a start.