api test: Fix a test to use Account.copyWith when it means that
The "auth headers sent by default" test below assumes that the
request is made with the email and API key of `eg.selfAccount`.
It happens to be true that `eg.account(user: eg.selfUser, …)`
will get that email, but that's a bit of an accident about
how eg.account works and what data is on eg.selfUser.

Instead of `eg.account`, use `eg.selfAccount.copyWith`, to make
the intention more explicit.