compose: Store controller on ComposeBox widget, not descendant
Now there's just one place in the code where the controller is
stored, instead of in both _StreamComposeBoxState and
_FixedDestinationComposeBoxState. Also, those two widgets are
simplified to be stateless.

One small functional change: if the compose box disappears and then
reappears -- because the self-user lost and gained posting
permission, or a DM recipient was deactivated then reactivated --
the compose box returns to its state before it disappeared, instead
of being re-initialized. This should be rare, but seems fine and
helpful when it happens; it's one fewer reason your compose progress
might get lost.