compose [nfc]: Bring _ComposeBoxContainer out of _ComposeBoxBody
Now, _ComposeBoxContainer is available as a nice place to put things
other than the "body" that (unlike the body) will want to consume
the left and right insets. Specifically, for #720, an error banner
and an overlaid progress indicator.

Next, we'll rename `_ComposeBoxContainer.child` to `body`. We'll
also go ahead and add a param for the error banner, redesign the
error banner that we currently have and are passing as `child` /
`body`, and pass the banner as that new param instead.