nav: Add bottom tabs and main menu
This is an initial implementation because some features were considered
out-of-scope as of now for #1035.

Compared to the Figma, we swapped the order of _ChannelsButton and
_DirectMessagesButton in the menu so they match their order on the
navigation bar. See:
  https://chat.zulip.org/#narrow/channel/48-mobile/topic/Buttons.20on.20the.20bottom.20tabs.20and.20main.20menu

We also added _CombinedFeedButton, using the same icon we have for
"Combined feed" on the web app's topleft sidebar, added a "Switch
account" button, and renamed "Streams" to "Channels".

For now, we do not aim to fully implement the app bar redesign; we keep a
simple one in this implementation. We maintain a 16px padding for the title
text on the app bar for both ChooseAccountPage (when it is at the root
navigation level) and HomePage (which is expected to be at the root level
all the time).

We allow user to navigate to the choose-account page from the loading screen
after a certain timeout.  This is helpful when loading the account data takes
a while, and that the user, for example, wants to try another account.
See: https://github.com/zulip/zulip-flutter/pull/1076#discussion_r1868262551

Signed-off-by: Zixuan James Li <zixuan@zulip.com>