internal_link test [nfc]: Give test groups distinct names
The organization of this code's test cases is a bit scattered;
partly because they were ported from tests written for the
legacy zulip-mobile RN app, which themselves accumulated in layers
over time.  (After all, one doesn't want to drop one of these tests
in a refactoring unless taking the time to be sure its substance is
covered by a surviving test.)

It's potentially confusing having several groups with the same name,
though.  Fix that by picking different names for all but one of them.

The name collision was originally noted by Zixuan in #1059.