msglist [nfc]: Add MessageListView to MessageListPageState
This will make it available in a long-press handler in the app bar
for a topic narrow. That long-press handler brings up the topic
action sheet, which will need a message ID to pass in the API
request for resolve/unresolve topic. (Not sure the API request
should *need* to require a message ID, but it does; shrug.)