emoji: Compute separator-plus-query just once per query
This is NFC except for performance: it saves us from having to
construct this string again for each emoji name.

At this stage there's a (much smaller) performance loss, too:
we now compute this string even when we'll never actually use it
because the query doesn't contain the separator.  But we'll soon
start checking for word-aligned matches even for those queries
where we'd accept a non-word-aligned match; once we do, we'll
need this string for all nonempty queries, for almost all names.