api [nfc]: Rename RealmEmojiItem.id to emojiCode
This way the binding encapsulates an important fact about this
bit of the API -- namely that these values are the same as the
"emoji code" values seen elsewhere in the API -- so that we don't
have to re-verify it when working elsewhere in the codebase and
using this class.

Also edit a doc comment in the model code to take advantage of
that understanding.