emoji [nfc]: Compare emoji base ordering with web's, in detail
I spent some time reverse-engineering in full detail the behavior web
has for building its list of possible emoji to present in the UI (and
so for the order that emoji results have when they don't differ on
any of the criteria used at the ranking step after applying a query).

The first output of that study was discovering #1113, fixed earlier
in this branch.  This records the rest.