icons: Explicitly sort in build-icon-font
The custom icon entries in "icons.dart" used to be generated/sorted
based on the order returned by Node for the related ".svg" files in
"assets\icons" directory. This default order behavior is different in
Node's implementation on different platforms (Windows vs. Linux or
macOS) thus resulting in inconsistent output of "icons.dart".
This PR adds an explicit sorting mechanism; so the output is
consistent across different platforms.

Motivated by:
  https://chat.zulip.org/#narrow/channel/516-mobile-dev-help/topic/CI.20failure.20related.20to.20Zulip.20icons.20-.20.23F889/near/1994820

Fixes: #1110