notif: Add vm:entry-point pragma on class, too
Fixes #1119.

With Flutter from main, in release mode, the app would no longer
show notifications on Android when it was in the background.
This error message appears in logcat:

```
DartVM   E  ERROR: To access 'package:zulip/notifications/receive.dart::NotificationService' from native code, it must be annotated.
            ERROR: See https://github.com/dart-lang/sdk/blob/master/runtime/docs/compiler/aot/entry_point_pragma.md
```

It looks like the relevant upstream commit is probably this one:
  https://github.com/dart-lang/sdk/commit/9b06e26620a081c89931e5618336ff23dc6083be

AFAICT those docs don't actually say that we should need to annotate
the class.  We already do annotate the static *method* on this class
which we ask native code to call, namely `_onBackgroundMessage`.
At least the error message is nice and clear, though.