autocomplete: Put best matches near input field. This commit reverses the list that was originally presented to the user while showing the typeahead menu.
This makes sense since on mobile its easier to click on options closer to the input box, i.e.
where your fingers are currently present, instead of pressing arrow keys on a keyboard which is
true on a desktop setup.

Hence we place the best matching options not at the top of the typeahead menu, but instead put
them at the bottom for better reachability and convenience of the user.

Fixes #1123.
Fixes #1121.